v2 of files removed and checkout epic started
