+++ enatega-multivendor-app/src/components/Main/ActiveOrdersAndSections/index.js
@@ -1,2 +1,2 @@
-import ActiveOrdersAndSections from './ActiveOrdersAndSectionsV2'
+import ActiveOrdersAndSections from './ActiveOrdersAndSections'
 export { ActiveOrdersAndSections }
