+++ enatega-multivendor-app/src/components/Main/Item/styles.js
@@ -17,22 +17,28 @@ const styles = (props = null) =>
         width: 0,
         height: verticalScale(0)
       },
-      borderRadius: scale(25),
-      height: scale(220),
+      borderColor: props != null ? props.borderColor : 'grey',
+      borderWidth: scale(2),
+      borderRadius: scale(8),
+      height: scale(270),
       width: '99%',
-      padding: scale(15),
+      //padding: scale(15),
 
       ...alignment.MBsmall
     },
     imageContainer: {
       position: 'relative',
+      zIndex: 1,
       alignItems: 'center',
+      width: '100%',
       height: '70%'
     },
     img: {
       width: '100%',
       height: '100%',
-      borderRadius: scale(18)
+
+      borderTopLeftRadius: scale(8),
+      borderTopRightRadius: scale(8)
     },
     overlayRestaurantContainer: {
       position: 'absolute',
