+++ enatega-multivendor-app/src/screens/CreateAccount/styles.js
@@ -12,14 +12,47 @@ const styles = (props = null) =>
       alignItems: 'center',
       backgroundColor: props !== null ? props.buttonText : 'transparent'
     },
+    container: {
+      flex: 1,
+      alignItems: 'center',
+      justifyContent: 'center',
+      backgroundColor: '#f3f0eb'
+    },
+    image: {
+      height: height * 0.5,
+      width: '100%',
+      marginBottom: scale(-20),
+
+      flex: 1,
+
+      shadowColor: '#f3f0eb',
+      shadowOffset: { width: 16, height: 8 },
+      shadowOpacity: 0.1,
+      shadowRadius: 19
+    },
+    image1: {
+      height: '100%',
+      width: '100%',
+      // marginBottom: scale(-10),
+      overflow: 'hidden'
+    },
     subContainer: {
+      alignSelf: 'center',
+      width: '100%',
+      flex: 1,
+      borderTopLeftRadius: scale(30),
+      borderTopRightRadius: scale(30),
+      backgroundColor: props != null ? props.themeBackground : '#FFF',
+      borderWidth: scale(1),
+      borderColor: props != null ? props.borderBottomColor : '#DAD6D6'
+    },
+    signupContainer: {
+      paddingVertical: scale(20),
       display: 'flex',
       alignSelf: 'center',
-      width: '80%',
-      paddingBottom: 10,
+      width: '100%',
       flex: 1,
-      justifyContent: 'center',
-      alignItems: 'center'
+      justifyContent: 'space-between'
     },
     whiteColor: {
       backgroundColor: props !== null ? props.buttonText : 'transparent'
@@ -48,9 +81,9 @@ const styles = (props = null) =>
       alignItems: 'center'
     },
     appleBtn: {
-      width: '100%',
-      height: height * 0.07,
-      fontSize: 20
+      width: '90%',
+      alignSelf: 'center',
+      height: height * 0.08
     },
     orText: {
       flexDirection: 'row',
@@ -60,6 +93,20 @@ const styles = (props = null) =>
       flex: 1,
       height: 1,
       backgroundColor: props !== null ? props.horizontalLine : '#9B9A9A'
+    },
+    guestButton: {
+      width: '90%',
+
+      alignSelf: 'center',
+      height: height * 0.08,
+      flexDirection: 'row',
+      alignItems: 'center',
+      ...alignment.PRlarge,
+      backgroundColor: 'transparent',
+      justifyContent: 'space-evenly',
+      borderRadius: scale(28),
+      borderWidth: scale(1),
+      borderBottomColor: '#000'
     }
   })
 export default styles
