+++ enatega-multivendor-app/src/screens/Login/Login.js
@@ -13,10 +13,10 @@ import styles from './styles'
 import Spinner from '../../components/Spinner/Spinner'
 import TextDefault from '../../components/Text/TextDefault/TextDefault'
 import { alignment } from '../../utils/alignment'
-import { FontAwesome } from '@expo/vector-icons'
+import { FontAwesome, Feather } from '@expo/vector-icons'
 import { useLogin } from './useLogin'
 import screenOptions from './screenOptions'
-import {useTranslation} from 'react-i18next'
+import { useTranslation } from 'react-i18next'
 
 function Login(props) {
   const {
@@ -35,7 +35,7 @@ function Login(props) {
     setShowPassword,
     checkEmailExist
   } = useLogin()
-    const {t} = useTranslation()
+  const { t } = useTranslation()
   useLayoutEffect(() => {
     props.navigation.setOptions(
       screenOptions({
@@ -61,17 +61,13 @@ function Login(props) {
           <View style={styles(currentTheme).mainContainer}>
             <View style={styles().subContainer}>
               <View style={styles().logoContainer}>
-                <Image
-                  source={require('../../../assets/login-icon.png')}
-                  style={styles().logoContainer}
-                />
+                <Feather name="mail" size={30} color="black" />
               </View>
               <View>
                 <TextDefault
                   H3
                   bolder
                   style={{
-                    textAlign: 'center',
                     ...alignment.MTlarge,
                     ...alignment.MBmedium
                   }}>
@@ -82,94 +78,94 @@ function Login(props) {
                     H5
                     bold
                     textColor={currentTheme.horizontalLine}
-                    style={{
-                      textAlign: 'center'
-                    }}>
-                    {registeredEmail
-                      ? t('signInWithEmail')
-                      : t('checkAccount')}
+                    style={{}}>
+                    {registeredEmail ? t('signInWithEmail') : t('checkAccount')}
                   </TextDefault>
                 )}
               </View>
               <View style={styles().form}>
                 <View>
-                  <TextInput
-                    placeholder={t('email')}
-                    style={[
-                      styles(currentTheme).textField,
-                      emailError !== null ? styles(currentTheme).errorInput : {}
-                    ]}
-                    placeholderTextColor={currentTheme.fontSecondColor}
-                    value={email}
-                    onChangeText={e => setEmail(e.toLowerCase().trim())}
-                  />
-                  {emailError !== null && (
-                    <TextDefault
-                      style={styles().error}
-                      bold
-                      textColor={currentTheme.textErrorColor}>
-                      {emailError}
-                    </TextDefault>
+                  <View>
+                    <TextInput
+                      placeholder={t('email')}
+                      style={[
+                        styles(currentTheme).textField,
+                        emailError !== null
+                          ? styles(currentTheme).errorInput
+                          : {}
+                      ]}
+                      placeholderTextColor={currentTheme.fontSecondColor}
+                      value={email}
+                      onChangeText={e => setEmail(e.toLowerCase().trim())}
+                    />
+                    {emailError !== null && (
+                      <TextDefault
+                        style={styles().error}
+                        bold
+                        textColor={currentTheme.textErrorColor}>
+                        {emailError}
+                      </TextDefault>
+                    )}
+                  </View>
+                  {registeredEmail && (
+                    <>
+                      <View style={styles().passwordField}>
+                        <TextInput
+                          secureTextEntry={showPassword}
+                          placeholder={t('password')}
+                          style={[
+                            styles(currentTheme).textField,
+                            styles().passwordInput,
+                            passwordError !== null
+                              ? styles(currentTheme).errorInput
+                              : {}
+                          ]}
+                          placeholderTextColor={currentTheme.fontSecondColor}
+                          value={password}
+                          onChangeText={e => setPassword(e)}
+                        />
+                        <FontAwesome
+                          onPress={() => setShowPassword(!showPassword)}
+                          name={showPassword ? 'eye' : 'eye-slash'}
+                          size={24}
+                          color={
+                            passwordError === null
+                              ? currentTheme.startColor
+                              : currentTheme.textErrorColor
+                          }
+                          style={[
+                            styles().eyeBtn,
+                            Platform.OS === 'android' && { marginTop: 40 }
+                          ]}
+                        />
+                      </View>
+                      {passwordError !== null && (
+                        <View>
+                          <TextDefault
+                            style={styles().error}
+                            bold
+                            textColor={currentTheme.textErrorColor}>
+                            {passwordError}
+                          </TextDefault>
+                        </View>
+                      )}
+                      <TouchableOpacity
+                        style={alignment.MBsmall}
+                        activeOpacity={0.7}
+                        onPress={() =>
+                          props.navigation.navigate('ForgotPassword', { email })
+                        }>
+                        <TextDefault
+                          textColor={currentTheme.main}
+                          style={alignment.MTsmall}
+                          bolder>
+                          {t('forgotPassword')}
+                        </TextDefault>
+                      </TouchableOpacity>
+                    </>
                   )}
                 </View>
-                {registeredEmail && (
-                  <>
-                    <View style={styles().passwordField}>
-                      <TextInput
-                        secureTextEntry={showPassword}
-                        placeholder={t('password')}
-                        style={[
-                          styles(currentTheme).textField,
-                          styles().passwordInput,
-                          passwordError !== null
-                            ? styles(currentTheme).errorInput
-                            : {}
-                        ]}
-                        placeholderTextColor={currentTheme.fontSecondColor}
-                        value={password}
-                        onChangeText={e => setPassword(e)}
-                      />
-                      <FontAwesome
-                        onPress={() => setShowPassword(!showPassword)}
-                        name={showPassword ? 'eye' : 'eye-slash'}
-                        size={24}
-                        color={
-                          passwordError === null
-                            ? currentTheme.startColor
-                            : currentTheme.textErrorColor
-                        }
-                        style={[
-                          styles().eyeBtn,
-                          Platform.OS === 'android' && { marginTop: 40 }
-                        ]}
-                      />
-                    </View>
-                    {passwordError !== null && (
-                      <View>
-                        <TextDefault
-                          style={styles().error}
-                          bold
-                          textColor={currentTheme.textErrorColor}>
-                          {passwordError}
-                        </TextDefault>
-                      </View>
-                    )}
-                    <TouchableOpacity
-                      style={alignment.MBxSmall}
-                      activeOpacity={0.7}
-                      onPress={() =>
-                        props.navigation.navigate('ForgotPassword', { email })
-                      }>
-                      <TextDefault
-                        textColor={currentTheme.buttonBackgroundPink}
-                        style={alignment.MTsmall}
-                        bold>
-                        {t('forgotPassword')}
-                      </TextDefault>
-                    </TouchableOpacity>
-                  </>
-                )}
-                <View style={styles().marginTop10}>
+                <View style={{ marginBottom: 20 }}>
                   <TouchableOpacity
                     onPress={() =>
                       registeredEmail
@@ -177,10 +173,10 @@ function Login(props) {
                         : checkEmailExist(email)
                     }
                     activeOpacity={0.7}
-                    style={styles().btn}>
+                    style={styles(currentTheme).btn}>
                     <TextDefault
                       H4
-                      textColor={currentTheme.buttonTextPink}
+                      textColor={currentTheme.black}
                       style={alignment.MLsmall}
                       bold>
                       {loading || loginLoading ? (
