+++ enatega-multivendor-app/src/screens/Login/styles copy.js
@@ -19,10 +19,8 @@ const styles = (props = null) =>
       ...alignment.PTlarge
     },
     subContainer: {
-      width: '90%',
-      height: '100%',
-      flex: 1,
-      flexDirection: 'column'
+      width: '85%',
+      height: '100%'
     },
     alignItemsCenter: {
       alignItems: 'center'
@@ -46,12 +44,14 @@ const styles = (props = null) =>
     appleBtn: {
       width: '100%'
     },
-    logoContainer: {},
+    logoContainer: {
+      width: scale(120),
+      height: scale(130),
+      alignSelf: 'center'
+    },
     form: {
       width: '100%',
-      flex: 1,
-      flexDirection: 'column',
-      justifyContent: 'space-between'
+      ...alignment.MTlarge
     },
     textField: {
       borderColor: props !== null ? props.borderColor : '#efefef',
@@ -66,7 +66,7 @@ const styles = (props = null) =>
       },
       shadowOpacity: 0.25,
       shadowRadius: 3.84,
-      elevation: 1,
+      elevation: 5,
       ...alignment.MTlarge
     },
     errorInput: {
@@ -91,12 +91,12 @@ const styles = (props = null) =>
     btn: {
       position: 'relative',
       // marginTop: '55%',
-      width: '100%',
+      width: '70%',
       alignItems: 'center',
-      backgroundColor: props !== null ? props.main : '#',
+      backgroundColor: props !== null ? props.black : '#000',
       alignSelf: 'center',
       padding: scale(15),
-      borderRadius: scale(28)
+      borderRadius: scale(10)
     },
     headerLeftIcon: {
       ...alignment.PLsmall
