+++ enatega-multivendor-app/src/screens/Menu/styles.js
