+++ enatega-multivendor-app/src/screens/Restaurant/styles.js
