+++ enatega-multivendor-app/src/screens/SelectLocation/AddNewAddress.js
@@ -1,39 +1,29 @@
 import React, {
   useState,
   useContext,
+  useCallback,
   useLayoutEffect,
-  useEffect,
-  useCallback
+  useRef,
+  useEffect
 } from 'react'
-import {
-  View,
-  Text,
-  TouchableOpacity,
-  StatusBar,
-  Linking,
-  TextInput
-} from 'react-native'
+import { View, Text, TouchableOpacity, TextInput } from 'react-native'
 import { LocationContext } from '../../context/Location'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
-import MapView, { PROVIDER_GOOGLE } from 'react-native-maps'
+
 import { theme } from '../../utils/themeColors'
 import TextDefault from '../../components/Text/TextDefault/TextDefault'
 import styles from './styles'
 import ThemeContext from '../../ui/ThemeContext/ThemeContext'
-import screenOptions from './screenOptions'
-import { useNavigation } from '@react-navigation/native'
-import { useLocation } from '../../ui/hooks'
-import { FlashMessage } from '../../ui/FlashMessage/FlashMessage'
-import { mapStyle } from '../../utils/mapStyle'
 import CustomMarker from '../../assets/SVG/imageComponents/CustomMarker'
-import analytics from '../../utils/analytics'
 import { customMapStyle } from '../../utils/customMapStyles'
 import { useTranslation } from 'react-i18next'
 import SearchModal from '../../components/Address/SearchModal'
 import { Feather } from '@expo/vector-icons'
 import ModalDropdown from '../../components/Picker/ModalDropdown'
-import { MaterialIcons } from '@expo/vector-icons'
 import { fetchAddressFromCoordinates } from '../../utils/geocoding'
+import { useNavigation } from '@react-navigation/native'
+import MapView from './MapView'
+import screenOptions from './screenOptions'
 
 const LATITUDE = 33.699265
 const LONGITUDE = 72.974575
@@ -41,67 +31,28 @@ const LATITUDE_DELTA = 40
 const LONGITUDE_DELTA = 40
 
 export default function AddNewAddress(props) {
-  const [searchModalVisible, setSearchModalVisible] = useState(false)
-  const [selectedCity, setSelectedCity] = useState('')
-  const [customAddress, setCustomAddress] = useState('')
-  const [selectedSearchAddress, setSelectedSearchAddress] = useState('')
-  const Analytics = analytics()
+  const { longitude, latitude } = props.route.params?.location || {}
+  console.log('AddNewAddress', latitude, longitude)
+  // city, latitude, longitude
 
-  const { t } = useTranslation()
-  const { longitude, latitude } = props.route.params || {}
+  const [searchModalVisible, setSearchModalVisible] = useState()
+  const [cityModalVisible, setCityModalVisible] = useState(false)
+  const [selectedValue, setSelectedValue] = useState({
+    city: '',
+    address: '',
+    latitude: '',
+    longitude: ''
+  })
+  const { setLocation } = useContext(LocationContext)
+  const mapRef = useRef()
+
   const themeContext = useContext(ThemeContext)
   const currentTheme = theme[themeContext.ThemeValue]
-  const navigation = useNavigation()
   const inset = useSafeAreaInsets()
-  const { getCurrentLocation, getLocationPermission } = useLocation()
-  const { setLocation } = useContext(LocationContext)
-  const [label, setLabel] = useState(
-    longitude && latitude ? t('currentLocation') : t('selectedLocation')
-  )
-  const [modalVisible, setModalVisible] = useState(false)
-  const [selectedValue, setSelectedValue] = useState(() => {
-    return props.route.params && props.route.params.city
-      ? props.route.params.city.name
-      : null
-  })
+  const navigation = useNavigation()
 
-  const openModal = () => {
-    setModalVisible(true)
-  }
+  const { t } = useTranslation()
 
-  const closeModal = () => {
-    setModalVisible(false)
-  }
-
-  const onItemPress1 = item => {
-    setSelectedValue(item.name) // Assuming your city object has a 'name' property
-    closeModal()
-  }
-
-  const onItemPress = item => {
-    onValueChange(item.value)
-    closeModal()
-  }
-  const openSearchModal = () => {
-    setSearchModalVisible(true)
-  }
-
-  const [coordinates, setCoordinates] = useState({
-    latitude: latitude || LATITUDE,
-    longitude: longitude || LONGITUDE,
-    latitudeDelta: latitude ? 0.003 : LATITUDE_DELTA,
-    longitudeDelta: longitude ? 0.003 : LONGITUDE_DELTA
-  })
-
-  useEffect(() => {
-    async function Track() {
-      await Analytics.track(Analytics.events.NAVIGATE_TO_SELECTLOCATION)
-    }
-    Track()
-    setCurrentLocation()
-  }, [])
-
-  let mapRef = null
   useLayoutEffect(() => {
     navigation.setOptions(
       screenOptions({
@@ -109,138 +60,59 @@ export default function AddNewAddress(props) {
         fontColor: currentTheme.fontMainColor,
         backColor: currentTheme.white,
         iconColor: currentTheme.black,
-        lineColor: currentTheme.lightHorizontalLine,
-        handleLocationWithoutCoords
+        lineColor: currentTheme.lightHorizontalLine
       })
     )
   }, [])
 
-  StatusBar.setBarStyle('dark-content')
-
-  const setCurrentLocation = async () => {
-    // Check if longitude and latitude are available in props
-    if (longitude && latitude) {
-      console.log('Inside setCurrentLocation', longitude, latitude)
-      // Set the current location coordinates on the map
-      mapRef.fitToCoordinates([
-        {
-          latitude,
-          longitude
-        }
-      ])
-
-      // Set the label to 'currentLocation'
-      setLabel('currentLocation')
-
-      // Set the current location address in the TextInput
-      setCustomAddress('') // Clear previous custom address if any
-      setSelectedSearchAddress('') // Clear previous selected search address if any
-
-      try {
-        // Now, fetch the address based on the obtained coordinates
-        const { formattedAddress, city } = await fetchAddressFromCoordinates(
-          latitude,
-          longitude
-        )
-
-        // Update the customAddress state with the current location address
-        setSelectedValue(city)
-        setCustomAddress(formattedAddress)
-      } catch (error) {
-        // Handle error fetching address
-        console.error('Error fetching address:', error.message)
-      }
-    }
+  const onSelectCity = item => {
+    console.log(' ', item)
+    setSelectedValue(item.name)
+    setCoordinates({ longitude: item.longitude, latitude: item.latitude })
+    setCityModalVisible(false)
   }
 
-  const handleLocationWithoutCoords = async () => {
-    console.log('inside Else in current LOcation')
-    // If no coordinates are available, use the logic to get the current location as before
-    const { status, canAskAgain } = await getLocationPermission()
-    if (status !== 'granted' && !canAskAgain) {
-      FlashMessage({
-        message: t('locationPermissionMessage'),
-        onPress: async () => {
-          await Linking.openSettings()
-        }
-      })
-      return
-    }
-
-    const { error, coords, message } = await getCurrentLocation()
-    if (error) {
-      FlashMessage({
-        message
-      })
-      return
-    }
-
-    // Set the current location coordinates on the map
-    mapRef.fitToCoordinates([
-      {
-        latitude: coords.latitude,
-        longitude: coords.longitude
-      }
-    ])
-
-    // Set the label to 'currentLocation'
-    setLabel('currentLocation')
-
-    // Set the current location address in the TextInput
-    setCustomAddress('') // Clear previous custom address if any
-    setSelectedSearchAddress('') // Clear previous selected search address if any
-
-    try {
-      // Now, fetch the address based on the obtained coordinates
-      const { formattedAddress, city } = await fetchAddressFromCoordinates(
-        coords.latitude,
-        coords.longitude
-      )
-
-      // Update the customAddress state with the current location address
-      console.log('fetch the city based on the obtained coordinates', city)
-      setSelectedValue(city)
-      setCustomAddress(formattedAddress)
-    } catch (error) {
-      // Handle error fetching address
-      console.error('Error fetching address:', error.message)
-    }
-  }
-
-  const onSelectSearchAddress = (address, coordinates) => {
-    console.log('address', address)
-    mapRef.fitToCoordinates([
-      {
-        latitude: coordinates.lat,
-        longitude: coordinates.lng
-      }
-    ])
-    setCustomAddress(address)
-    setSelectedSearchAddress(address)
-    setSearchModalVisible(false)
-    console.log('onSelectSearchAddress', coordinates, typeof mapRef)
-  }
-
-  const onRegionChangeComplete = coordinates => {
-    console.log('onRegionChangeComplete')
-    setCoordinates({
-      ...coordinates
-    })
-  }
-
-  const onPanDrag = () => {
-    console.log('onPanDrag')
-    setLabel(t('selectedLocation'))
-    setSearchModalVisible(false)
-  }
-
-  const onSelectLocation = () => {
-    setLocation({
-      label: selectedSearchAddress,
-      deliveryAddress: customAddress,
+  const onRegionChangeComplete = useCallback(async coordinates => {
+    console.log('onRegionChangeComplete', coordinates)
+    const response = await fetchAddressFromCoordinates(
+      coordinates.latitude,
+      coordinates.longitude
+    )
+    setSelectedValue({
+      city: response.city,
+      address: response.formattedAddress,
       latitude: coordinates.latitude,
       longitude: coordinates.longitude
     })
+    // populate city, address info based on coordinates
+  })
+
+  const onSelectSearch = location => {
+    setCoordinates(location)
+  }
+
+  const setCoordinates = useCallback(location => {
+    mapRef.current.fitToCoordinates([
+      {
+        latitude: location.latitude,
+        longitude: location.longitude
+      }
+    ])
+  })
+
+  useEffect(() => {
+    console.log('useEffect')
+    setCoordinates({ longitude, latitude })
+  }, [])
+
+  const onSelectLocation = () => {
+    setLocation({
+      label: 'Location',
+      deliveryAddress: selectedValue.address,
+      latitude: selectedValue.latitude,
+      longitude: selectedValue.longitude
+    })
+    navigation.navigate('Main')
   }
 
   return (
@@ -248,20 +120,19 @@ export default function AddNewAddress(props) {
       <View style={styles().flex}>
         <View style={[styles().mapView, { height: '55%' }]}>
           <MapView
-            ref={ref => {
-              mapRef = ref
+            ref={mapRef}
+            initialRegion={{
+              latitude: LATITUDE,
+              longitude: LONGITUDE,
+              latitudeDelta: LATITUDE_DELTA,
+              longitudeDelta: LONGITUDE_DELTA
             }}
-            initialRegion={coordinates}
-            // region={coordinates}
-            style={{ flex: 1 }}
-            provider={PROVIDER_GOOGLE}
-            showsTraffic={false}
-            maxZoomLevel={15}
             customMapStyle={
-              themeContext.ThemeValue === 'Dark' ? mapStyle : customMapStyle
+              themeContext.ThemeValue === 'Dark'
+                ? customMapStyle
+                : customMapStyle
             }
             onRegionChangeComplete={onRegionChangeComplete}
-            onPanDrag={onPanDrag}
           />
           <View style={styles().mainContainer}>
             <CustomMarker
@@ -281,37 +152,34 @@ export default function AddNewAddress(props) {
             style={styles().addressHeading}>
             {t('addAddress')}
           </TextDefault>
-          <View style={styles().dropdownContainer}>
-            <TouchableOpacity style={styles().button1} onPress={openModal}>
-              {selectedValue ? (
-                <Text>{selectedValue}</Text>
-              ) : (
-                <Text style={styles().placeholder}>Select City</Text>
-              )}
-              <Feather
-                name="chevron-down"
-                size={18}
-                color="black"
-                style={styles().icon1}
-              />
-            </TouchableOpacity>
-            <ModalDropdown
-              visible={modalVisible}
-              onItemPress={onItemPress1}
-              onClose={closeModal}
-            />
-          </View>
+          <CityModal
+            setCityModalVisible={setCityModalVisible}
+            selectedValue={selectedValue.city}
+            cityModalVisible={cityModalVisible}
+            onSelect={onSelectCity}
+          />
 
           <View style={[styles(currentTheme).textInput]}>
+            <TouchableOpacity onPress={() => setSearchModalVisible(true)}>
+              <Text
+                style={{
+                  color: currentTheme.buttonText,
+
+                  overflow: 'scroll'
+                }}>
+                {selectedValue.address || 'Address'}
+              </Text>
+            </TouchableOpacity>
+          </View>
+          {/* <View style={[styles(currentTheme).textInput]}>
             <TextInput
               style={[styles().flex, { color: currentTheme.buttonText }]}
               placeholder="Address"
               placeholderTextColor={currentTheme.placeholderText}
-              value={customAddress}
-              //onFocus={openSearchModal}
-              onChangeText={text => setCustomAddress(text)}
+              value={selectedValue.address}
+              onChangeText={text => {}}
             />
-            <TouchableOpacity onPress={openSearchModal}>
+            <TouchableOpacity onPress={() => setSearchModalVisible(true)}>
               <MaterialIcons
                 name="edit"
                 size={18}
@@ -319,7 +187,7 @@ export default function AddNewAddress(props) {
                 style={{ marginRight: 10, marginLeft: 10 }}
               />
             </TouchableOpacity>
-          </View>
+          </View> */}
           <TouchableOpacity
             activeOpacity={0.7}
             style={styles(currentTheme).emptyButton}
@@ -328,17 +196,64 @@ export default function AddNewAddress(props) {
               Save
             </TextDefault>
           </TouchableOpacity>
-
-          {searchModalVisible ? (
-            <SearchModal
-              visible={searchModalVisible}
-              onClose={() => setSearchModalVisible(false)}
-              onSubmit={onSelectSearchAddress}
-            />
-          ) : null}
+          <SearchModal
+            visible={searchModalVisible}
+            onClose={() => setSearchModalVisible(false)}
+            onSubmit={(...props) => {
+              onSelectSearch({
+                latitude: props[1].lat,
+                longitude: props[1].lng
+              })
+              setSearchModalVisible(false)
+              console.log('onSearchItemSelect', props[1].lat, props[1].lng)
+            }}
+          />
         </View>
         <View style={{ paddingBottom: inset.bottom }} />
       </View>
     </>
   )
 }
+
+const CityModal = React.memo(
+  function CityModal({
+    setCityModalVisible,
+    selectedValue,
+    cityModalVisible,
+    onSelect
+  }) {
+    return (
+      <View style={styles().dropdownContainer}>
+        <TouchableOpacity
+          style={styles().button1}
+          onPress={() => {
+            setCityModalVisible(true)
+          }}>
+          {selectedValue && <Text>{selectedValue}</Text>}
+          {!selectedValue && (
+            <Text style={styles().placeholder}>Select City</Text>
+          )}
+          <Feather
+            name="chevron-down"
+            size={18}
+            color="black"
+            style={styles().icon1}
+          />
+        </TouchableOpacity>
+        <ModalDropdown
+          visible={cityModalVisible}
+          onItemPress={onSelect}
+          onClose={() => {
+            setCityModalVisible(false)
+          }}
+        />
+      </View>
+    )
+  },
+  (prevProps, nextProps) => {
+    return (
+      prevProps.cityModalVisible === nextProps.cityModalVisible &&
+      prevProps.selectedValue === nextProps.selectedValue
+    )
+  }
+)
