+++ enatega-multivendor-app/src/ui/FdSocialBtn/FdEmailBtn/FdEmailBtn.js
@@ -8,10 +8,10 @@ import { theme } from '../../../utils/themeColors'
 import ThemeContext from '../../ThemeContext/ThemeContext'
 import { alignment } from '../../../utils/alignment'
 import TextDefault from '../../../components/Text/TextDefault/TextDefault'
-import {useTranslation} from 'react-i18next'
+import { useTranslation } from 'react-i18next'
 
 const FdEmailBtn = props => {
-  const {t} = useTranslation()
+  const { t } = useTranslation()
   const themeContext = useContext(ThemeContext)
   const currentTheme = theme[themeContext.ThemeValue]
   return (
@@ -20,18 +20,18 @@ const FdEmailBtn = props => {
       style={styles().mainContainer}
       onPress={props.onPress}>
       {props.loadingIcon ? (
-        <Spinner backColor="rgba(0,0,0,0.1)" spinnerColor={'#FFF'} />
+        <Spinner backColor="rgba(0,0,0,0.1)" spinnerColor={'#000'} />
       ) : (
         <>
           <MaterialIcons
             style={styles().marginLeft5}
             name="mail-outline"
             size={scale(18)}
-            color={currentTheme.white}
+            color={currentTheme.black}
           />
           <TextDefault
             H4
-            textColor={currentTheme.buttonTextPink}
+            textColor={currentTheme.black}
             style={alignment.MLsmall}
             bold>
             {t('ContinueWithEmail')}
