+++ enatega-multivendor-app/src/ui/FdSocialBtn/FdEmailBtn/styles copy.js
@@ -1,22 +1,16 @@
 import { Dimensions, StyleSheet } from 'react-native'
 import { alignment } from '../../../utils/alignment'
-import { scale } from '../../../utils/scaling'
 const { height } = Dimensions.get('window')
 
 const styles = (props = null) =>
   StyleSheet.create({
     mainContainer: {
-      width: '90%',
-      alignSelf: 'center',
-      height: height * 0.08,
+      height: height * 0.07,
+      backgroundColor: props !== null ? props.buttonBackgroundPink : '#000000',
       flexDirection: 'row',
       alignItems: 'center',
-      ...alignment.PRlarge,
-      backgroundColor: 'transparent',
-      justifyContent: 'space-evenly',
-      borderRadius: scale(28),
-      borderWidth: scale(1),
-      borderBottomColor: '#000'
+      borderRadius: 10,
+      ...alignment.PLsmall
     },
     marginLeft5: {
       ...alignment.MLsmall
