+++ enatega-multivendor-app/src/ui/FdSocialBtn/FdGoogleBtn/FdGoogleBtn.js
@@ -6,10 +6,10 @@ import { scale } from '../../../utils/scaling'
 import Spinner from '../../../components/Spinner/Spinner'
 import TextDefault from '../../../components/Text/TextDefault/TextDefault'
 import { alignment } from '../../../utils/alignment'
-import {useTranslation} from 'react-i18next'
+import { useTranslation } from 'react-i18next'
 
 const FdGoogleBtn = props => {
-  const {t} = useTranslation()
+  const { t } = useTranslation()
   return (
     <TouchableOpacity
       activeOpacity={0.7}
@@ -24,9 +24,9 @@ const FdGoogleBtn = props => {
             style={styles.marginLeft5}
             name="google"
             size={scale(18)}
-            color="#FFF"
+            color="#000"
           />
-          <TextDefault H4 textColor="#FFF" style={alignment.MLsmall} bold>
+          <TextDefault H4 textColor="#000" style={alignment.MLsmall} bold>
             {t('ContinueWithGoogle')}
           </TextDefault>
         </>
