+++ enatega-multivendor-app/src/ui/FdSocialBtn/FdGoogleBtn/styles copy.js
@@ -1,23 +1,17 @@
 import { Dimensions } from 'react-native'
 import { alignment } from '../../../utils/alignment'
 import { textStyles } from '../../../utils/textStyles'
-import { scale } from '../../../utils/scaling'
 const { height } = Dimensions.get('window')
 
 export default {
   mainContainer: {
-    width: '90%',
-
-    alignSelf: 'center',
-    height: height * 0.08,
+    width: '100%',
+    height: height * 0.07,
+    backgroundColor: '#b71c1c',
     flexDirection: 'row',
     alignItems: 'center',
-    ...alignment.PRlarge,
-    backgroundColor: 'transparent',
-    justifyContent: 'space-evenly',
-    borderRadius: scale(28),
-    borderWidth: scale(1),
-    borderBottomColor: '#000'
+    borderRadius: 10,
+    ...alignment.PRlarge
   },
   marginLeft5: {
     ...alignment.MLsmall
@@ -28,6 +22,6 @@ export default {
   textStyle: {
     ...textStyles.Bold,
     ...textStyles.Normal,
-    color: '#000'
+    color: '#FFF'
   }
 }
