+++ enatega-multivendor-app/src/ui/FdSocialBtn/FdGoogleBtn/styles.js
@@ -1,17 +1,23 @@
 import { Dimensions } from 'react-native'
 import { alignment } from '../../../utils/alignment'
 import { textStyles } from '../../../utils/textStyles'
+import { scale } from '../../../utils/scaling'
 const { height } = Dimensions.get('window')
 
 export default {
   mainContainer: {
-    width: '100%',
-    height: height * 0.07,
-    backgroundColor: '#b71c1c',
+    width: '90%',
+
+    alignSelf: 'center',
+    height: height * 0.08,
     flexDirection: 'row',
     alignItems: 'center',
-    borderRadius: 10,
-    ...alignment.PRlarge
+    ...alignment.PRlarge,
+    backgroundColor: 'transparent',
+    justifyContent: 'space-evenly',
+    borderRadius: scale(28),
+    borderWidth: scale(1),
+    borderBottomColor: '#000'
   },
   marginLeft5: {
     ...alignment.MLsmall
@@ -22,6 +28,6 @@ export default {
   textStyle: {
     ...textStyles.Bold,
     ...textStyles.Normal,
-    color: '#FFF'
+    color: '#000'
   }
 }
